requirements.txt
python-telegram-bot==21.4
nest_asyncio==1.6.0
requests==2.32.3